requirements.txt
python-telegram-bot==20.7
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.1
python-dotenv==1.0.0
dnspython==2.4.2
flask==3.0.0